requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-flatpickr==1.0.1
dodgy==0.1.9
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.2
PyYAML==5.1.2
requirements-detector==0.6
setoptconf==0.2.0
whitenoise==4.1.4
